pkg/tuple: remove tuple and replace with CtMap
To avoid ambiguity the pkg/tuple is now removed and replaced with the
CtMap Keys and Values. As pkg/tuple was used in pkg/maps/nat, there is a
new interface created in `pkg/maps/ctmap` called NatMap to avoid import
cycles from the pkg/maps/ctmap and pkg/maps/nat.

Signed-off-by: André Martins <andre@cilium.io>