eni: Support masquerading
This adds ENI support to the masquerading logic. For external SNAT, the
destination address exclusion is derived from the VPC's primary CIDR.

For now, the list of interfaces to masquerade on must be specified with
--egress-masquerade-interfaces with an interface selector. This is required
because filtering all local ENIs with a static rule is not possible yet. This
will be resolved once we switch to BPF based masquerading.

Signed-off-by: Thomas Graf <thomas@cilium.io>